requirements.txt
ament_copyright==0.12.14
ament_flake8==0.12.14
ament_index_python==1.4.0
ament_pep257==0.12.14
cv_bridge==3.2.1
message_filters==4.3.11
numpy==1.21.5
pyserial==3.5
pytest==6.2.5
PyYAML==5.4.1
PyYAML==6.0.3
rclpy==3.3.19
scipy==1.8.0
sensor_msgs==4.9.0
setuptools==58.2.0
setuptools==59.6.0
ultralytics==8.4.0
